requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
black==26.1.0
certifi==2026.2.25
cfgv==3.5.0
click==8.3.1
distlib==0.4.0
exceptiongroup==1.3.1
fastapi==0.133.1
filelock==3.24.3
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.16
idna==3.11
iniconfig==2.3.0
isort==8.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
pre_commit==4.5.1
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
python-discovery==1.0.0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
starlette==0.52.1
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
uvloop==0.22.1
virtualenv==21.0.0
watchfiles==1.1.1
websockets==16.0
